Move account endpoint serializers to file
